refactor(qwik/a2ui): remove unused mock stream helpers and enhance documentation in ElmA2ui stories
